# Screenkernal scrolling code optimizations