@johanberntsson 2022-10-11
Thanks for the PR. From what we understand, this saves ~1/10 s when scrolling the screen a full 25 lines. This may not be worth the cost of 100 bytes to us.

With every change or addition we make, we weigh the benefit to the cost, and to a large extent, the cost is measured in bytes. We’re using about 62.5 KB of the RAM in the C64, to make sure the interpreter needs to load game data from disk as seldomly as possible. We also have a limitation that only RAM up to $cfff can be used for dynamic memory - the portion of a game file that works as RAM rather than ROM. As the interpreter becomes bigger, it sometimes passes another 512-byte border, giving us 512 bytes less for game data in total, as well as 512 bytes less for dynamic memory. This limits which games are playable on the C64.