@0ldSkull 2022-10-12
That is a fair consideration, and I appreciate the additional information.
In reviewing my timing measurements again, I think that these optimizations are also probably not strictly required for smooth scrolling as well.  Outside of Ozmoo, having this part of the code be as fast as possible was desirable.  But it doesn't look as critical with the way I'm currently scheduling the movement and scrolling activities.

Here is some detail for how each change affects the cycles and bytes saved/gained:
```
opt	cycles  bytes
 1	  -904   +35
 2	  -503   -19
 3	 -1500   +48
 4	  -448   +34
```
It may be that 2 could be potentially worthwhile in terms of cost (since it presents a savings of both cycles and memory), though the gain is somewhat minimal.  For 3 independently the values would probably be half of the incremental effects shown.  (4 is applicable only after 1.)

If you want to consider e.g. just the second optimization independently then I can do some further testing and evaluation with that.  Or we can drop it if that is best, and I can still work on smooth scrolling without the additional optimized data movement.